+++ apps/ymir-sdl3/src/app/app.hpp
@@ -83,6 +83,8 @@ private:
     void ScanROMCarts();
     void LoadRecommendedCartridge();
 
+    void ApplyGameSpecificConfiguration();
+
     void LoadSaveStates();
     void ClearSaveStates();
     void LoadSaveStateSlot(size_t slot);